Going to try to fix all warnings in the application. Header files are already ok.

+++ Windows-Wrapper/ListItem.cpp
@@ -1,7 +1,7 @@
 #include "ListItem.h"
 #include "ListControl.h"
 
-ListItem::ListItem(uint32_t id, const std::string& value)
+ListItem::ListItem(uint64_t id, const std::string& value)
 	:
 	Id(id),
 	Value(value),
